8 byte floats for pika

+++ src/tokens/FloatToken.java
@@ -3,21 +3,21 @@ package tokens;
 import inputHandler.TextLocation;
 
 public class FloatToken extends TokenImp {
-	protected Float value;
+	protected Double value;
 	
 	protected FloatToken(TextLocation location, String lexeme) {
 		super(location, lexeme);
 	}
-	protected void setValue(Float value) {
+	protected void setValue(Double value) {
 		this.value = value;
 	}
-	public Float getValue() {
+	public Double getValue() {
 		return value;
 	}
 	
 	public static FloatToken make(TextLocation location, String lexeme) {
 		FloatToken result = new FloatToken(location, lexeme);
-		result.setValue(Float.parseFloat(lexeme));
+		result.setValue(Double.parseDouble(lexeme));
 		return result;
 	}
